changed roomName -> gameId & add initGuesser
